backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
django-cors-headers==4.0.0
openai==0.27.8
python-dotenv==1.0.0
